+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2 [unreleased]
 
 - Support `lat`, `lng`, and `lon`
+- Added labels and tooltips
 
 ## 0.0.1